Reduce chance of name collisions with concurrent harvest workers
Reduces the chances of collisions for harvesters concurrently processing
datasets where the name is the same.  By default, a sequential number is
appended by CKAN in order to prevent name collisions on the unique name field.
However, if several workers attempt to process a dataset of the same
name, they can end up using the same name, which will cause an SQL
duplicate error.  By appending a random hex string instead, the likelihood of a
collision is greatly reduced.  In the event a collision still does
happen, an error is logged and the worker continues processing other datasets
from the queue.  Should fix #95.